Add typeclass constraints - rearrange modules to resolve circular dependencies.
