+++ hie-plugin-api/hie-plugin-api.cabal
@@ -18,6 +18,7 @@ flag pedantic
 library
   exposed-modules:
                        Haskell.Ide.Engine.ArtifactMap
+                       Haskell.Ide.Engine.GhcModuleCache
                        Haskell.Ide.Engine.IdeFunctions
                        Haskell.Ide.Engine.ModuleCache
                        Haskell.Ide.Engine.Monad
@@ -25,6 +26,7 @@ library
                        Haskell.Ide.Engine.MonadTypes
                        Haskell.Ide.Engine.MultiThreadState
                        Haskell.Ide.Engine.PluginDescriptor
+                       Haskell.Ide.Engine.PluginsIdeMonads
                        Haskell.Ide.Engine.PluginTypes
                        Haskell.Ide.Engine.PluginUtils
   build-depends:       base >= 4.9 && < 5
